Refactor new-dao
